Behavioral Platform Implementation codes/Utilities/requirements.txt
opencv-contrib-python
numpy
pandas
pyserial
pyautogui
pysimplegui
matplotlib